+++ .sequelizerc
@@ -1,0 +1,5 @@
+const path = require('path');
+
+module.exports = {
+    config : path.resolve('config','database.js')
+}
