+++ config/database.js
@@ -1,0 +1,9 @@
+const config = {
+    username : 'root',
+    password : '140908',
+    database : 'barberhairdb',
+    host : 'localhost',
+    dialect: 'mysql'
+}
+
+module.exports = config
