--- controllers/userController.js
@@ -1,17 +1,0 @@
-const userServices = require('../services/userServices')
-
-module.exports = {
-
-    findAll(req,res,next){
-        console.log('Entrou na Controller')
-        res.send(userServices.fyndAll())
-    },
-
-    cadastro(req,res,next){
-        res.render('cadastroUsuario.ejs')
-    },
-
-    cadastro2(req,res,next){
-        res.send(userServices.insertUsers)
-    }
-}
