+++ public/stylesheets/style.css
@@ -1,11 +1,44 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  scroll-behavior: smooth;
+}
+
 body {
   padding: 0px;
   font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
+  background: #F4F4F4;
+}
+ 
+footer {
+  height: 100px;
+  
 }
 
-a {
-  color: #00B7FF;
+.formInput {
+height: 30px ;
+margin-left: 30px;
 }
+
+
+a {
+  color: black !important;
+}
+
+section.cadastroForms{
+  
+  margin-top: 20px;
+  overflow: hidden;
+  padding-bottom: 10px;
+  padding-top: 20px;
+  padding-left: 20px;
+  padding-right: 20px;
+  border-radius: 30 30 30 30;
+  width: 100%;
+  
+}
+
 .imageFundo {
    background-image: url(../images/Fundo.jpg);
    padding: 100px;
@@ -13,21 +46,26 @@ a {
    background-size: cover ;
    background-position: center;
    display: inline-block;
+   padding-top: 0px;
+   height: 170px;
 
  }
-
- .cadastroSalao{
-  padding-left: 20px;
-  padding-right: 20px;
-  border-radius: 30 30 30 30
-   
- }
-
 
  .fontLabel{
   font-family: Kristen ITC, Arial,sans-serif;
-  font-size: 14px;
+  font-size: 12px;
   font-weight: 550;
+  
+ }
+
+ .fontLabelIndex{
+  font-family: Kristen ITC, Arial,sans-serif;
+  font-size: 17px;
+  font-weight: 550;
+  /*margin-left: 15%;*/
+  color: #a56e00;
+  text-align:center; 
+  align-content: center; 
   
  }
 
@@ -48,15 +86,68 @@ a {
   border-bottom-left-radius: 20px;
 }
 
- .cadastroSalaoDiv{
-   margin-left: 10px;
- }
 
- form{
-  border-radius: 25px
- }
-  footer {
-    height: 100px;
-  }
+.botoesnav{
+  font-family: Kristen ITC, Arial,sans-serif;
+  background-color: rgba(255, 255, 255, 0.5) !important;
+  -webkit-border-radius: 20px;
+  -moz-border-radius: 20px;
+  border-radius: 20px;
+  padding: 5px 10px 5px 10px;
   
+}
+
+.divnotificacoes{
+  border-radius: 20px;
+  margin-top: 20px;
+  border: 1px solid #252525;
+  width: 160px;
+  height: 100px;
+  display: inline-block;
+}
+
+
+.divnotificacoes input{
+  margin-left: 30px;
+}
+
+.divnotificacoes label{
+  margin-left: 15px;
+}
+
+.divFoto{
+  border-radius: 20px;
+  margin-top: 20px;
+  border: 1px solid #252525;
+  width: 120px;
+  height: 160px;
+  display: inline-block;
+}
+
+.botaoAgendar{
+  background-color: rgba(165,42,42, 0.5);
+  border-radius:20px;  
+  display: inline-block; 
+  padding-left: 20px; 
+  padding-right: 20px;
+}
+
+.hrIndex{
+  border-color:black;
+  /*margin-left: 10% ; */
+  box-sizing:border-box;
+  width:40%;  
+}
+
+.imgDestaque{
+  height: 15vh;
+}
+
+.divDestaque{
+  /*margin-left: 5%; 
+  margin-right: 5%;
+*/
+}
+
+
