+++ routes/cadastroSalao.js
@@ -1,11 +1,9 @@
 var express = require('express');
-const cadastroSalaoController = require('../controllers/cadastroSalaoController');
 var router = express.Router();
 var CadastroSalaoController = require('../controllers/cadastroSalaoController')
 
 router.get('/',CadastroSalaoController.cadastro)
 router.post('save',CadastroSalaoController.save)
-
 
 
 module.exports = router;
