+++ views/cadastroUsuario.ejs
@@ -1,10 +1,169 @@
-<%- include('./template/cabecalho'); %> 
+<%- include('./template/cabecalho'); %>
+
+  <main class="container-lg col-md-9" style="background-color: lightgray; border-radius: 20px;">
+
+    <section class="cadastroForms" id="sectionFormCadastro">
+
+      <form action="/cadastroUser/save" method="post" style="overflow: hidden;">
+
+        <div class="container" style=" width: 50%; float: left; margin-left: 95px;">
+          <div class="row mb-3 " style="padding-top: 20px;">
+            <label for="inputNome" class=" col-sm-1 col-form-label fontLabel">Nome</label>
+            <div class="col-sm-8">
+              <input type="text" class="form-control formInput" id="inputNome">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputCpfCnpj" class="col-sm-1 col-form-label fontLabel">CPF</label>
+            <div class="col-sm-8">
+              <input type="text" class="form-control formInput" id="inputCpfCnpj">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputEndereco" class="col-sm-1 col-form-label fontLabel">Endereço</label>
+            <div class="col-sm-8">
+              <input type="text" class="form-control formInput" id="inputEndereco">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputCidade" class="col-sm-1 col-form-label fontLabel">Cidade</label>
+            <div class="col-sm-8">
+              <input type="text" class="form-control formInput" id="inputCidade">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputEstado" class="col-sm-1 col-form-label fontLabel">Estado</label>
+            <div class="col-sm-5">
+              <input type="text" class="form-control formInput" id="inputEstado">
+            </div>
+          </div>
+
+
+          <div class="row mb-3">
+            <label for="inputCep" class="col-sm-1 col-form-label fontLabel">Cep</label>
+            <div class="col-sm-5">
+              <input type="text" class="form-control formInput" id="inputCep">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputEmail" class="col-sm-1 col-form-label fontLabel">Email</label>
+            <div class="col-sm-8">
+              <input type="email" class="form-control formInput" id="inputEmail">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputContato" class="col-sm-1 col-form-label fontLabel">Celular</label>
+            <div class="col-sm-8">
+              <input type="text" class="form-control formInput" id="inputContato">
+            </div>
+          </div>
+
+          <div class="row mb-3">
+            <label for="inputSenha" class="col-sm-1 col-form-label fontLabel">Senha</label>
+            <div class="col-sm-8">
+              <input type="password" class="form-control formInput" id="inputSenha">
+            </div>
+          </div>
+
+          <div class="row mb-3" style="padding-bottom: 20px;">
+            <label for="inputConfSenha" class="col-sm-1 col-form-label fontLabel">Confirmação</label>
+            <div class="col-sm-8">
+              <input type="password" class="form-control formInput" id="inputConfSenha">
+            </div>
+          </div>
+
+        </div>
+        <div class="container" style="margin-left: 10px; width: 40%; float: right;">
+          
+          <div class="row mb-3 divFoto" >
+            
+          </div>
+          <br>
+          <div class="row mb-3 divnotificacoes" >
+            <label class="fontLabel" style="margin-top: 10px; ">Notificações</label>
+            <br>
+            <input type="checkbox" id="sms" name="sms" value="S"><label class="fontLabel" for="sms">SMS</label>
+            <br>
+            <input type="checkbox" id="email" name="email" value="S"><label for="email">Email</label><br>
+          </div>
+          <br>
+
+
+        </div>
+
+       
+        </div>
 
 
 
+      </form>
+    </section>
 
+    <section class="cadastroForms" id="sectionUltimosAgend">
+      <div class="container" style="padding-bottom: 20px;">
+        <label class="fontLabelTable">Lista dos Ultimos Agendamentos</label>
 
+        <table class="table table-hover " style="width: 100%; text-align: center;">
+          <tr>
+            <th scope="col">#</th>
+            <th scope="col">Data</th>
+            <th scope="col">Salão</th>
+            <th scope="col">Procedimento Agendado</th>
+            <th scope="col">Valor</th>
+            <th scope="col">Efetivado</th>
+            <th scope="col">Agendar Novamente</th>
+          </tr>
+          <tbody>
+            <tr>
+              <th scope="row">1</th>
+              <td>10/02/2020</td>
+              <td>Zé Fuinha</td>
+              <td>Corte</td>
+              <td>R$100,00</td>
+              <td>Sim</td>
+              <td style="padding-bottom: 5px;padding-top: 5px;">
+                <div class="botaoAgendar">
+                  <a href="/users/agendamentos" class="nav-link">Agendar</a>
+                </div>
+              </td>
+            </tr>
+            <tr>
+              <th scope="row">2</th>
+              <td>10/02/2020</td>
+              <td>Zé Fuinha</td>
+              <td>Corte</td>
+              <td>R$100,00</td>
+              <td>Não</td>
+              <td style="padding-bottom: 5px;padding-top: 5px;">
+                <div class="botaoAgendar">
+                  <a href="/users/agendamentos" class="nav-link">Agendar</a>
+                </div>
+              </td>
+            </tr>
+            <tr>
+              <th scope="row">3</th>
+              <td>10/02/2020</td>
+              <td>Zé Fuinha</td>
+              <td>Corte</td>
+              <td>R$100,00</td>
+              <td>Sim</td>
+              <td  style="padding-bottom: 5px;padding-top: 5px;">
+                <div class="botaoAgendar">
+                  <a href="/users/agendamentos" class="nav-link">Agendar</a>
+                </div>
+              </td>
+            </tr>
 
+          </tbody>
+        </table>
+      </div>
+    </section>
+  </main>
 
-
-<%- include('./template/rodape'); %> 
+  <%- include('./template/rodape'); %>
