+++ views/index copy.ejs
@@ -1,0 +1,91 @@
+<%- include('./template/cabecalho'); %>
+  <main class='container' style="max-width: 90%; margin-left: 8%;">
+    <section class="indexSection" id="indexSection">
+
+      <div class="" style=" margin-top: 100px;">
+        <label class=" fontLabelIndex">Serviços em Destaque</label>
+        <hr class="hrIndex" />
+      </div>
+      <div class="card-deck">
+        <div class="col-12 col-md-6 col-lg-3 mt-3">
+          <div class="card" style="width: 18rem;">
+            <img class="card-img-top imgDestaque" src="../images/corteDesenhado.png" alt="Card image cap">
+            <div class="card-body">
+              <h5 class="card-title">Card title</h5>
+              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.</p>
+            </div>
+            <div class="card-footer">
+              <small class="text-muted">Last updated 3 mins ago</small>
+            </div>
+          </div>
+        </div>
+
+        <div class="col-12 col-md-6 col-lg-3 mt-3">
+          <div class="card" style="width: 18rem;">
+            <img class="card-img-top imgDestaque" src="../images/Barba.jpg" alt="Card image cap" width="">
+            <div class="card-body">
+              <h5 class="card-title">Card title</h5>
+              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.</p>
+            </div>
+            <div class="card-footer">
+              <small class="text-muted">Last updated 3 mins ago</small>
+            </div>
+          </div>
+        </div>
+
+        <div class="col-12 col-md-6 col-lg-3 mt-3">
+          <div class="card" style="width: 18rem;">
+            <img class="card-img-top imgDestaque" src="../images/Barba.jpg" alt="Card image cap" width="">
+            <div class="card-body">
+              <h5 class="card-title">Card title</h5>
+              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.</p>
+            </div>
+            <div class="card-footer">
+              <small class="text-muted">Last updated 3 mins ago</small>
+            </div>
+          </div>
+        </div>
+
+        <div class="col-12 col-md-6 col-lg-3 mt-3">
+          <div class="card" style="width: 18rem;">
+            <img class="card-img-top imgDestaque" src="../images/corte2.jpg" alt="Card image cap" width="">
+            <div class="card-body">
+              <h5 class="card-title">Card title</h5>
+              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.</p>
+            </div>
+            <div class="card-footer">
+              <small class="text-muted">Last updated 3 mins ago</small>
+            </div>
+          </div>
+        </div>
+
+        <div class="col-12 col-md-6 col-lg-3 mt-3">
+          <div class="card" style="width: 18rem;">
+            <img class="card-img-top imgDestaque" src="../images/Barba.jpg" alt="Card image cap" width="">
+            <div class="card-body">
+              <h5 class="card-title">Card title</h5>
+              <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional
+                content. This content is a little bit longer.</p>
+            </div>
+            <div class="card-footer">
+              <small class="text-muted">Last updated 3 mins ago</small>
+            </div>
+          </div>
+        </div>
+
+      </div>
+    </section>
+
+  </main>
+
+
+
+
+
+
+
+  <%- include('./template/rodape'); %>
