+++ views/login.ejs
@@ -1,0 +1,54 @@
+<%- include('./template/cabecalho'); %>
+
+
+
+    <main>
+        <section>
+            <div id="layoutAuthentication">
+                <div id="layoutAuthentication_content">
+                    <main>
+                        <div class="container">
+                            <div class="row justify-content-center">
+                                <div class="col-lg-5">
+                                   <div class="card shadow-lg border-0 rounded-lg mt-5">
+                                    <div class="row " style="justify-content: center; padding-top: 100px;">
+                                        <img src="../images/user.png" class="imgLogin">
+                                    </div>
+                                        <div class="card-body">
+                                            <form>
+                                                <div class="form-group">
+                                                    <label class="small mb-1" for="inputEmailAddress">Email</label>
+                                                    <input class="form-control py-4" id="inputEmailAddress" type="email" placeholder="abc@gmailcom" />
+                                                </div>
+                                                <div class="form-group">
+                                                    <label class="small mb-1" for="inputPassword">Senha</label>
+                                                    <input class="form-control py-4" id="inputPassword" type="password" placeholder="Senha" />
+                                                </div>
+                                                <div class="form-group">
+                                                    <div class="custom-control custom-checkbox">
+                                                        <input class="custom-control-input" id="rememberPasswordCheck" type="checkbox" />
+                                                        <label class="custom-control-label" for="rememberPasswordCheck">Lembrar Senha</label>
+                                                    </div>
+                                                </div>
+                                                <div class="form-group d-flex align-items-center justify-content-between mt-4 mb-0">
+                                                    <a class="small" href="password.html">Esqueceu a senha?</a>
+                                                    <a class="btn btn-primary" href="index.html">Login</a>
+                                                </div>
+                                            </form>
+                                        </div>
+                                        <div class="card-footer text-center">
+                                            <div class="small"><a href="/users/create">Ainda Não tem cadastro! Registre-se!</a></div>
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </main>
+                </div>
+        </section>
+    </main>
+
+
+
+
+    <%- include('./template/rodape'); %>
