+++ views/template/cabecalho.ejs
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home</title>
+    <title>BarberHair</title>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.7.1/css/all.css" integrity="sha384-fnmOCqbTlWIlj8LyTjo7mOUStjsKC4pOpQbqyi7RrhN7udi9RwhKkMHpvLbHG9Sr" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
     <link rel='stylesheet' href='/stylesheets/style.css' />
@@ -12,6 +12,23 @@
 
   <body>
     <header>
-      <div class="imageFundo col-md-12">
-      </div>
+      <nav class="navbar navbar-expand-lg navbar-light bg-light imageFundo col-md-12">
+        <div class="container  col-md-12" style="padding-top: 10px;">
+          
+          <ul class="navbar-nav botoesnav"  >
+            <img src="../images/userbtn.png">
+            <li class="nav-item ">
+              <a href="/users/create" class="nav-link ">Cadastre-se</a>
+            </li>
+          </ul>
+          <ul class="navbar-nav ml-auto  botoesnav">
+          <li class="nav-item">
+            <a href="/cadastroSalao" class="nav-link">Cadastre seu Salão</a>
+          </li>
+
+        </ul>
+        </div> 
+      </nav>
+ 
+    </nav>
     </header>
