+++ views/template/rodape.ejs
@@ -8,7 +8,7 @@
       Portal de Agendamento &copy; todos os direitos reservados
     </div>
     <div>
-      Desenvolvido por Patrick/Jorge/Nathalia
+      Desenvolvido por Patrick/Jorge
     </div>
   </footer>
 </body>
